+++ gamechangerml/configs/__init__.py
@@ -10,4 +10,4 @@ from .d2v_config import D2VConfig
 from .validation_config import ValidationConfig
 from .bert_summary_config import BertSummConfig
 from .topics_config import TopicsConfig
-from .path_config import PathConfig
+from .redis_config import RedisConfig
