+++ gamechangerml/configs/redis_config.py
@@ -1,0 +1,5 @@
+from os import environ
+
+class RedisConfig:
+    HOST = environ.get("REDIS_HOST", default="localhost")
+    PORT = environ.get("REDIS_PORT", default=6379)
